Collectable
Created a collectable that is collected upon a player colliding with it

Created predefined locations that the collectables are instantiated in

Collectables increase a score on the player's screens

Current bug: Connecting clients pick up collectables delayed. This could be due to the custom collision that I have implemented. Potential fix is to revert the custom collision and attempt to use Unity's collisions and use that to feed information to the StateAuthority