Began making lobby
Created a scriptable object containing the game data. This data only must be read by the host, so it is perfect for saving settings.

Created a simple lobby where all four players are presented in front of the camera

Created setting allowing players to change how many collectables are needed to win

TODO:
- Make the lobby change the scene to the game
- Make the lobby transfer all the players from the lobby to the game
- Allow players to change the color of their character using the buttons inside of the lobby
- Make the main menu take players to the lobby

After this has been done, I will be working on improving the main menu, adding audio and settings.

Furthermore, I will be trying to improve the visuals of the game this week.